refactor(tui): wiki もタブごとのディレクトリへ揃える
issues / time_entries と同じ構成にして、タブ関連のファイルの置き場所を統一する。
wiki には専用の modal がないため wiki_tab.py のみを移動した。
